Renamed 'container' folder in algebra-plugin to 'definitions'
