+++ plugins/algebra/include/detray/eigen/include/plugins/algebra/eigen_container.hpp
@@ -1,5 +1,5 @@
 
-#include "algebra/containers/eigen.hpp"
+#include "algebra/definitions/eigen.hpp"
 
 #define ALGEBRA_PLUGIN eigen
